+++ app/models/document_model.py
@@ -4,6 +4,7 @@ from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey
 from sqlalchemy.orm import relationship
 from app.database import Base
 from app.config import get_config
+from app.models.tag_models import DocumentTag
 
 cfg = get_config()
 
@@ -34,6 +35,9 @@ class Document(Base):
         "User", back_populates="user_documents", lazy="joined")
     document_collection = relationship(
         "Collection", back_populates="collection_documents", lazy="joined")
+    document_tags = relationship(
+        "Tag", secondary=DocumentTag.__table__, back_populates="tag_documents",
+        lazy="joined")
 
     def __init__(self, user_id: int, collection_id: int, document_name: str,
                  filename: str, filesize: int, mimetype: str,
@@ -60,6 +64,12 @@ class Document(Base):
         if self.thumbnail_filename:
             return cfg.THUMBNAILS_BASE_URL + self.thumbnail_filename
 
+    @property
+    def _document_tags(self) -> list:
+        if self.document_tags:
+            return [x.value for x in self.document_tags]
+        return []
+
     def to_dict(self):
         return {
             "id": self.id,
@@ -76,4 +86,5 @@ class Document(Base):
 
             "thumbnail_url": self.thumbnail_url,
             "comments_count": self.comments_count,
+            "document_tags": self._document_tags,
         }