+++ app/repository.py
@@ -1,3 +1,12 @@
+"""
+This module defines the Repository class, which provides a unified
+interface for managing CRUD operations and caching for SQLAlchemy
+models, using an async session for database interactions and Redis
+for caching. It includes methods for checking existence, inserting,
+selecting, updating, deleting, counting, and summing models, along
+with transaction management through commit and rollback.
+"""
+
 from typing import List, Type, Union
 from redis import Redis
 from sqlalchemy.ext.asyncio import AsyncSession
@@ -7,28 +16,40 @@ from app.managers.cache_manager import CacheManager
 
 
 class Repository:
-    """Manages CRUD operations and caching for SQLAlchemy models."""
+    """
+    A repository class for managing CRUD operations and caching for
+    SQLAlchemy models.
+    """
 
     def __init__(self, session: AsyncSession, cache: Redis,
                  entity_class: Type[DeclarativeBase]):
-        """Initializes a repository for specific SQLAlchemy model."""
+        """
+        Initializes the repository with an async session, Redis cache,
+        and the SQLAlchemy model class to manage.
+        """
         self.entity_manager = EntityManager(session)
         self.cache_manager = CacheManager(cache)
         self.entity_class = entity_class
 
     async def exists(self, **kwargs) -> bool:
-        """Checks if an entity exists with the given criteria."""
+        """
+        Checks if a SQLAlchemy model matching the given criteria exists
+        in the database.
+        """
         return await self.entity_manager.exists(self.entity_class, **kwargs)
 
     async def insert(self, entity: DeclarativeBase, commit: bool = True):
-        """Inserts an entity and optionally caches it."""
+        """
+        Inserts a new SQLAlchemy model into the database, with optional
+        immediate transaction commit.
+        """
         await self.entity_manager.insert(entity, commit=commit)
 
-        if self.entity_class._cacheable and commit:
-            await self.cache_manager.set(entity)
-
     async def select(self, **kwargs) -> Union[DeclarativeBase, None]:
-        """Retrieves an entity by id or other criteria."""
+        """
+        Retrieves a SQLAlchemy model based on the provided criteria
+        or ID, using cache if available.
+        """
         entity_id, entity = kwargs.get(ID), None
 
         if self.entity_class._cacheable and entity_id:
@@ -48,7 +69,10 @@ class Repository:
         return entity
 
     async def select_all(self, **kwargs) -> List[DeclarativeBase]:
-        """Retrieves all entities matching the given criteria."""
+        """
+        Retrieves all SQLAlchemy models that match the given criteria,
+        with optional caching.
+        """
         entities = await self.entity_manager.select_all(
             self.entity_class, **kwargs)
 
@@ -59,35 +83,48 @@ class Repository:
         return entities
 
     async def update(self, entity: DeclarativeBase, commit: bool = True):
-        """Updates an entity and manages its cache status."""
+        """
+        Updates an existing SQLAlchemy model in the database, with
+        optional immediate transaction commit.
+        """
         await self.entity_manager.update(entity, commit=commit)
 
         if self.entity_class._cacheable:
-            if commit:
-                await self.cache_manager.set(entity)
-            else:
-                await self.cache_manager.delete(entity)
+            await self.cache_manager.delete(entity)
 
     async def delete(self, entity: DeclarativeBase, commit: bool = True):
-        """Deletes an entity and manages its cache status."""
+        """
+        Deletes a SQLAlchemy model from the database, with optional
+        immediate transaction commit.
+        """
         await self.entity_manager.delete(entity, commit=commit)
 
         if self.entity_class._cacheable:
             await self.cache_manager.delete(entity)
 
     async def count_all(self, **kwargs) -> int:
-        """Counts all entities matching the given criteria."""
+        """
+        Counts the number of SQLAlchemy models that match the given
+        criteria.
+        """
         return await self.entity_manager.count_all(self.entity_class, **kwargs)
 
     async def sum_all(self, column_name: str, **kwargs) -> int:
-        """Sums a column's values for entities matching the criteria."""
+        """
+        Calculates the sum of a specific column for all SQLAlchemy
+        models matching the criteria.
+        """
         return await self.entity_manager.sum_all(
             self.entity_class, column_name, **kwargs)
 
     async def commit(self):
-        """Commits the current transaction."""
+        """
+        Commits the current transaction to the database.
+        """
         await self.entity_manager.commit()
 
     async def rollback(self):
-        """Rolls back the current transaction."""
+        """
+        Rolls back the current transaction in case of issues.
+        """
         await self.entity_manager.rollback()